moves state transition logic into its own module
